refactor(b00t-pgduck): split type-mapping fns into a lib target
main.rs was bin-only, so the repo pre-push hook's `cargo test --lib -p
b00t-pgduck` failed 'no library targets found' and blocked every push touching
this crate. Move the two pure, IO-free type-mapping fns (arrow_type_to_pg_type,
duckdb_type_str_to_pg_type) into src/lib.rs, add a [lib] target, and import them
back into main.rs. No behaviour change — byte-identical fn bodies.

⚠️ cargo check could not run here (host rust toolchain is broken — same reason
the pre-commit/pre-push hooks fail); verified by inspection: mechanical
function move, call sites unchanged, deps shared with the bin target.
